explicitly cast to UInt to avoid warning in VC9
git-svn-id: svn://hevc.kw.bbc.co.uk/svn/jctvc-tmuc/branches/HM-4.1-dev@1676 ce97f6c2-8f86-0410-b44d-af84b13ec34f

+++ source/Lib/TLibEncoder/TEncBinCoderCABACCounter.cpp
@@ -54,13 +54,13 @@ TEncBinCABACCounter::~TEncBinCABACCounter()
 
 Void TEncBinCABACCounter::finish()
 {
-  m_pcTComBitIf->write(0, m_fracBits >> 15);
+  m_pcTComBitIf->write(0, UInt(m_fracBits >> 15) );
   m_fracBits &= 32767;
 }
 
 UInt TEncBinCABACCounter::getNumWrittenBits()
 {
-  return m_pcTComBitIf->getNumberOfWrittenBits() + ( m_fracBits >> 15 );
+  return m_pcTComBitIf->getNumberOfWrittenBits() + UInt( m_fracBits >> 15 );
 }
 
 /**
